Feat: create module HelperEncrypt
