+++ decrypted.txt
@@ -1,1 +1,1 @@
-good afternoon
+good night
